specs: add commands format
Users of Hermes devices communicate with the device through commands,
which are handled by a driver.

Introduce some of the commands and their request and response formats.

Some commands are passed verbatim to the device and as such use the same
specification for both the user <-> driver and host <-> device
interfaces.

Other commands use the XDMA interface for the host <-> device interface,
but provide a different, Hermes-specific user <-> driver interface.
Translation between these two interfaces is handled by the Hermes
driver. These commands are marked in the document.