drivers: add xdma support to hermes driver
We export some XDMA functions so that the Hermes drivers can call them,
thus enabling XDMA through the Hermes driver. This creates the
/dev/xdmaX_{c2h,h2c}_Y devices for Hermes devices. Note that Hermes
devices are *not* bound to the XDMA driver, only to the Hermes driver.

Since the XDMA probe function enables the device, we do not need to do
it in the Hermes driver anymore.

The Hermes driver now depends on the XDMA driver, which is managed
automatically.