drivers/xdma: create exported functions to get dma char devices
These are the /dev/xdmaX_{c2h,h2c}_Y devices that can be used to do
DMAs. We will need them in the Hermes driver to initiate DMAs from the
Hermes driver.

For now, only return the first DMA channel (if it exists). In time, we
can create more elaborate heuristics to select an idle channel.