driver: create functions to select dma device
To use XDMA, a user needs to explicitly select a channel, which is
clearly not optimal. To deal with this, introduce functions to select a
channel.

Initially, we only return the first DMA channel (if it exists), but in
time we can improve it to find and return an idle channel.