driver: add support for the Write and Read commands
The Write (0x10) and Read (0x11) commands can be used to move data
between host and device. These are just wrappers around XDMA, with two
benefits:
1. The same interface as all other Hermes commands
2. It hides hardware details from the user, as it chooses a DMA channel
automatically.

They are implemented as ioctls, where the command request is passed in
the first 32 bytes of the ioctl's third parameter (argp). The command
response is written to the subsequent 16 bytes. The ioctl 'request'
parameter is ignored.