driver: create module parameter to enable/disable XDMA channel devices
It is a security issue to have the XDMA channels exposed, as anyone can
read/write to arbitrary locations.

Now that we support data transfer through /dev/hermesX, we can close
this hole.

Instead of completely removing these devices, we instead make them
optional, as they can still be useful for debugging. A module parameter
is introduced to enable/disable the device creation (defaults to off).

Update the unit tests to skip the XDMA tests if it can't find the
device.