specs: add BAR0 register for command requests/responses
For every eBPF engine (EHENG), we have one set of command
request/response registers (starting at 0x1000) and one one-byte control
register (starting at 0x2000).

The user must write to the command request register, then write to the
corresponding control register, which starts the command. When the
command finishes, the hardware writes to the command response registers
and clears the control register.

Note that the above does not apply to the Write/Read commands, which are
handled by the hermes/xdma driver.

Fixes #16