driver: do not create char devices for DMA channels
This is the first commit (of two) that removes the XDMA char devices for
each channel. Unlike the other char devices, which we completely
removed, we need to keep a few functions related to this device (those
that actually do the DMA).

This commit introduces the xdma_channel struct, which keeps only the
relevant members from the xdma_cdev struct, and converts the code to use
this struct. The next commit will do a cleanup to remove all the new
dead code.

Now that the devices aren't being created, we need to also get rid of
the unit tests, but we will add new tests later after we have added the
Hermes char device.

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>