driver: add support for program/data transfer
This patch introduces support for downloading an eBPF program to a
program slot and to transfer data to/from a data slot.

An eBPF program can be downloaded to the device through an ioctl,
passing the program and its length as the argp parameter of the ioctl.

Data slots are written/read through the write()/read() syscalls.

Every file descriptor of the Hermes char device will have associated to
it a program slot and data slot, which are allocated the first time we
issue the ioctl or write syscalls. Slots are currently allocated in
software but might later be done on hardware.

In the future, the write() syscall will also trigger the execution of
the corresponding program. For that reason, we disallow writes without
a previous ioctl.

If a user needs to run multiple programs, then the user should use
multiple file descriptors (using a second ioctl will overwrite the
current program slot). Note that this prevents effective "program
chaining", that is, using the unmodified output of a program as the
input of another.

Signed-off-by: Martin Oliveira <martin.oliveira@eideticom.com>