driver: drop support for older kernels
With #ifdef's for kernels 3.5, 3.16, 4.12, 4.13, the code was getting a
bit of a maintenance pain. Popular distros (Ubuntu 18.04/20.04 and
CentOS/RHEL 8) already use kernels above 4.13, so the chance of those
macros being useful is diminished.

A notable exception is RHEL/CentOS 7, which is still on 3.10 and will
thus no longer be supported.

We keep only one LINUX_VERSION_CODE macro, for kernels < 5.1.0, as that
is required to support RHEL/CentOS 8.