driver: introduce channel_interrupts_mask()
The enable/disable functions were nearly identical, except that enable
was writing to the W1S register and and the disable was writing to the
W1C. The introduced channel_interrupts_mask() writes to the regular
version of the register, so it can be shared by the enable/disable
functions.

No functional changes.