driver: move channel interrupt enable/disable calls
A call to irq_teardown() is always preceeded by

	channel_interrupts_disable(xdev);
	read_interrupts(xdev);

Also irq_teardown() is currently just a wrapper around
xdma_irq_teardown() (though this will change in the future with the
addition of eBPF interrupts). So we move the above couple of calls into
the begining of xdma_irq_teardown().

Similarly, a call to irq_setup() is always followed by

	channel_interrupts_enable(xdev);
	read_interrupts(xdev);

And irq_setup() is just a wrapper around xdma_irq_setup(), so move those
two calls to the end of xdma_irq_setup().

No functional changes.