driver: move channel interrupt enable/disable calls
A call to xdma_irq_teardown() is always preceeded by

	channel_interrupts_disable(xdev);
	read_interrupts(xdev);

So we move the above couple of calls into the begining of
xdma_irq_teardown().

Similarly, a call to irq_setup() is always followed by

	channel_interrupts_enable(xdev);
	read_interrupts(xdev);

So we move those two calls to the end of xdma_irq_setup().

No functional changes.