unit_test, example: use valid eBPF programs
So far the driver tests and examples were using dummy eBPF programs,
which was fine because we weren't executing them. This is no longer the
case, so we need a valid eBPF program.

So use a very simple program, that consits of just two instructions:

r0 = 0 # opcode 0xb7
ret    # opcode 0x95