bugfix: Fix ehcmddone race-condition.
When offloading a program after the first one, the state of Hermes causes a race condition to occour.
As such, Hermes gives control back to the application before the result of the eBPF execution is transferred to the host.
Ultimately, it will be the response of the previous execution that the application receives.

Setting the ehcmddone bit host-side resolves this.

Resolves #42